+++ VisualChatGPTStudioShared/Utils/Enums.cs
@@ -1,4 +1,4 @@
-﻿namespace JeffPires.VisualChatGPTStudio.Utils
+namespace JeffPires.VisualChatGPTStudio.Utils
 {
     /// <summary>
     /// Represents the different types of requests that can be used.
@@ -17,5 +17,15 @@
         Replace,
         InsertBefore,
         InsertAfter
+    }
+
+    /// <summary>
+    /// Specifies the available options for selecting a Copilot model.
+    /// </summary>
+    public enum CopilotModelOption
+    {
+        Default = 0,
+        Completion = 1,
+        Specific = 2
     }
 }